refactor(*) Move current time utility from compiler to engine
