+++ packages/ergo-compiler/test/module.js
@@ -34,7 +34,6 @@ describe('Module', () => {
             index.Logger.should.not.be.null;
             index.ScriptManager.should.not.be.null;
             index.TypeException.should.not.be.null;
-            index.Util.should.not.be.null;
             index.FileLoader.should.not.be.null;
             index.ErgoLoader.should.not.be.null;
             index.version.should.not.be.null;
