+++ packages/ergo-engine/lib/util.js
